requirements.txt
asgiref==3.5.2
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
contourpy==1.0.6
copydetect==0.4.3
cycler==0.11.0
Django==3.1.6
djangorestframework==3.12.2
et-xmlfile==1.1.0
fonttools==4.38.0
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
mccabe==0.7.0
memory-profiler==0.61.0
multimetric==1.3.0
numpy==1.23.5
openai==0.25.0
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
pandas-stubs==1.5.2.221124
Pillow==9.3.0
psutil==5.9.4
pycodestyle==2.10.0
pydocstyle==6.1.1
pyflakes==3.0.1
Pygments==2.13.0
pylama==8.4.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.3
tqdm==4.64.1
types-pytz==2022.6.0.1
typing-extensions==4.4.0
urllib3==1.26.13
